Sanity refactor
