+++ src/test/create-all-templates.spec.js
@@ -1,14 +1,9 @@
-/**
- * Created by JP Strydom (JPStrydom) and  Muzikayise Flynn Buthelezi (zuluCoda) on 2018/01/27.
- * mfbproject.co.za - muzi@mfbproject.co.za
- */
-
 'use strict';
 
 const fs = require('fs');
 const path = require('path');
 
-const { getRootPath } = require('../generator-utils');
+const { getRootPath } = require('../generator-utilities');
 
 const createAllTemplates = require('../create-all-templates');
